upload-service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.0
fastapi==0.115.12
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.0
idna==3.10
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
